requirements.txt
simplepyble
pyyaml
noise